requirements.txt
requests
fastapi 
uvicorn 
azure-functions